fix uploadfile bugs

+++ composeApp/src/commonMain/kotlin/irancell/nwg/wfm/File.kt
@@ -1,7 +1,7 @@
 package irancell.nwg.wfm
 
 expect class File(path: String) {
-    fun sizeInMB(): Int
+    fun sizeInMB(): Double
     fun readBytes(): ByteArray
     fun extension(): String
     fun exists(): Boolean
